+++ src/HomePage.ts
@@ -1,28 +1,52 @@
-import Server from './server';
 import * as vscode from "vscode";
 import { HOMEPAGE } from './enum';
+import Server from './server';
 
 export default class HomePage {
 
   /**
-	 * Track the currently panel. Only allow a single panel to exist at a time.
-	 */
-	public static currentPanel: HomePage | undefined;
-  
+   * Track the currently panel. Only allow a single panel to exist at a time.
+   */
+  public static currentPanel: HomePage | undefined;
+
   private readonly _panel: vscode.WebviewPanel;
-	private readonly _extensionUri: vscode.Uri;
+  private readonly _extensionUri: vscode.Uri;
+  private _disposables: vscode.Disposable[] = [];
   public readonly _server: Server;
 
-  public static createOrShow(extensionUri: vscode.Uri, server: Server) {
-		const column = vscode.window.activeTextEditor
-			? vscode.window.activeTextEditor.viewColumn
-			: undefined;
+  private constructor(panel: vscode.WebviewPanel, extensionUri: vscode.Uri, server: Server) {
+    this._panel = panel;
+    this._extensionUri = extensionUri;
+    this._server = server;
 
-		// If we already have a panel, show it.
-		if (HomePage.currentPanel) {
-			HomePage.currentPanel._panel.reveal(column);
-			return;
-		}
+    // Set the webview's initial html content
+    this._update();
+
+    // Listen for when the panel is disposed
+    // This happens when the user closes the panel or when the panel is closed programmatically
+    this._panel.onDidDispose(this.dispose, null, this._disposables);
+
+    // Handle messages from the webview
+    this._panel.webview.onDidReceiveMessage(
+      message => {
+        switch (message.command) {
+          case 'startServer':
+            this._server.restartServer().then(this._update.bind(this))
+            return;
+        }
+      }, null, this._disposables);
+  }
+
+  public static createOrShow(extensionUri: vscode.Uri, server: Server) {
+    const column = vscode.window.activeTextEditor
+      ? vscode.window.activeTextEditor.viewColumn
+      : undefined;
+
+    // If we already have a panel, show it.
+    if (HomePage.currentPanel) {
+      HomePage.currentPanel._panel.reveal(column);
+      return;
+    }
 
     const panel = vscode.window.createWebviewPanel(
       HOMEPAGE, // Identifies the type of the webview. Used internally
@@ -31,47 +55,47 @@ export default class HomePage {
       {
         enableScripts: true, // Enable scripts in the webview
         // And restrict the webview to only loading content from our extension's `media` directory.
-          localResourceRoots: [vscode.Uri.joinPath(extensionUri, 'public')]
+        localResourceRoots: [vscode.Uri.joinPath(extensionUri, 'public')]
       }
     );
 
-		HomePage.currentPanel = new HomePage(panel, extensionUri, server);
+    HomePage.currentPanel = new HomePage(panel, extensionUri, server);
+  }
+
+  public dispose() {
+		HomePage.currentPanel = undefined;
+
+		// Clean up our resources
+		this._panel.dispose();
+
+		while (this._disposables.length) {
+			const x = this._disposables.pop();
+			if (x) {
+				x.dispose();
+			}
+		}
 	}
 
-  private constructor(panel: vscode.WebviewPanel, extensionUri: vscode.Uri, server: Server) {
-		this._panel = panel;
-		this._extensionUri = extensionUri;
-    this._server = server;
-		this._panel.webview.html = this.getWebviewContent(this._panel.webview);
+  private _update() {
+    const webview = this._panel.webview;
 
-		// Listen for when the panel is disposed
-		// This happens when the user closes the panel or when the panel is closed programmatically
-		this._panel.onDidDispose(() => {
-      HomePage.currentPanel = undefined;
-      this._panel.dispose()
-    }, null);
-
-    // Handle messages from the webview
-		this._panel.webview.onDidReceiveMessage(
-			message => {
-				switch (message.command) {
-					case 'startServer':
-              this._server.restartServer().then(() => {
-                this._panel.webview.html = this.getWebviewContent(this._panel.webview);
-              })
-						return;
-				}
-			},
-			null,
-		);
-	}
+    // Vary the webview's content based on where it is located in the editor.
+    switch (this._panel.viewColumn) {
+      case vscode.ViewColumn.Three:
+      case vscode.ViewColumn.Two:
+      case vscode.ViewColumn.One:
+      default:
+        this._panel.webview.html = this.getWebviewContent(webview);
+        return;
+    }
+  }
 
   public static revive(panel: vscode.WebviewPanel, extensionUri: vscode.Uri, server: Server) {
-		HomePage.currentPanel = new HomePage(panel, extensionUri, server);
-	}
+    HomePage.currentPanel = new HomePage(panel, extensionUri, server);
+  }
 
   private getWebviewContent = (webview: vscode.Webview) => {
-    
+
     const config = this._server?.mockServer?.config || {};
     const { port, host, base } = config;
     return `
@@ -101,19 +125,17 @@ export default class HomePage {
       </script>
     </head>
     <body style="margin: 0; padding: 0; height: 100vh; width: 100%">
-    ${
-      this._server?.mockServer?.server ?
-      `<iframe id="iframe-data"
+    ${this._server?.mockServer?.server ?
+        `<iframe id="iframe-data"
       style="width: 100%; height: 100%" 
-      src="http://${host}:${port}${base}?_fontSize=11px&_dataFontSize=0.9rem&_dataLineHeight=1.4"
+      src="http://${host}:${port}${base}?_fontSize=11px&_dataFontSize=0.8rem&_dataLineHeight=1.1"
       frameborder="0"></iframe>`
-      :
-      `<div id="start-server"><a onclick="startServer()">Click Here to Start the Mock Server</a></div>`
-    }
+        :
+        `<div id="start-server"><a onclick="startServer()">Click Here to Start the Mock Server</a></div>`
+      }
     </body>
     <script>
       function startServer() {
-        console.log('clicked!');
         vscode.postMessage({
           command: 'startServer',
         });
@@ -121,5 +143,5 @@ export default class HomePage {
     </script>
     </html>`
   }
-  
+
 }