+++ src/enum.ts
@@ -1,7 +1,7 @@
-export const GENERATE_ROUTES = "mockServer.generateRoutes";
+export const TRANSFORM_TO_MOCK_SERVER_DB = "mockServer.transformToMockServerDB";
+export const GENERATE_MOCK_FILES = "mockServer.generateMockFiles";
 export const START_SERVER = "mockServer.startServer";
 export const STOP_SERVER = "mockServer.stopServer";
-export const RESET_SERVER = "mockServer.resetServer";
 export const SWITCH_ENVIRONMENT = "mockServer.switchEnvironment";
 export const GET_DB_SNAPSHOT = "mockServer.getDbSnapShot";
 export const HOMEPAGE = "mockServer.homePage";